Some functions has been extracted form act_comm.cpp
