+++ src/act_comm.cpp
@@ -44,139 +44,6 @@ void do_gen_comm(CharData *ch, char *argument, int cmd, int subcmd);
 void do_pray_gods(CharData *ch, char *argument, int cmd, int subcmd);
 void do_ignore(CharData *ch, char *argument, int cmd, int subcmd);
 
-#define SIELENCE ("Вы немы, как рыба об лед.\r\n")
-#define SOUNDPROOF ("Стены заглушили ваши слова.\r\n")
-
-void do_say(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-
-//	create_charmice_stuff(ch, static_cast<ESkill>(atoi(argument)), 100);// г номер тестировал чармстаф
-
-	if (AFF_FLAGGED(ch, EAffect::kSilence)) {
-		SendMsgToChar(SIELENCE, ch);
-		return;
-	}
-
-	if (ch->IsFlagged(EPlrFlag::kDumbed)) {
-		SendMsgToChar("Вам запрещено обращаться к другим игрокам!\r\n", ch);
-		return;
-	}
-
-	if (!*argument)
-		SendMsgToChar("Вы задумались: \"Чего бы такого сказать?\"\r\n", ch);
-	else {
-		sprintf(buf, "$n сказал$g : '%s'", argument);
-
-// для возможности игнорирования теллов в клетку
-// пришлось изменить act в клетку на проход по клетке
-		for (const auto to : world[ch->in_room]->people) {
-			if (ch == to || ignores(to, ch, EIgnore::kSay)) {
-				continue;
-			}
-			act(buf, false, ch, nullptr, to, kToVict | DG_NO_TRIG | kToNotDeaf);
-		}
-
-		if (ch->IsFlagged(EPrf::kNoRepeat)) {
-			SendMsgToChar(OK, ch);
-		} else {
-			delete_doubledollar(argument);
-			sprintf(buf, "Вы сказали : '%s'\r\n", argument);
-			SendMsgToChar(buf, ch);
-		}
-		speech_mtrigger(ch, argument);
-		speech_wtrigger(ch, argument);
-	}
-}
-
-void do_gsay(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-	CharData *k;
-	struct FollowerType *f;
-
-	if (AFF_FLAGGED(ch, EAffect::kSilence)) {
-		SendMsgToChar(SIELENCE, ch);
-		return;
-	}
-
-	if (ch->IsFlagged(EPlrFlag::kDumbed)) {
-		SendMsgToChar("Вам запрещено обращаться к другим игрокам!\r\n", ch);
-		return;
-	}
-
-	skip_spaces(&argument);
-
-	if (!AFF_FLAGGED(ch, EAffect::kGroup)) {
-		SendMsgToChar("Вы не являетесь членом группы!\r\n", ch);
-		return;
-	}
-
-	if (!*argument) {
-		SendMsgToChar("О чем вы хотите сообщить своей группе?\r\n", ch);
-	} else {
-		if (ch->has_master()) {
-			k = ch->get_master();
-		} else {
-			k = ch;
-		}
-
-		sprintf(buf, "$n сообщил$g группе : '%s'", argument);
-
-		if (AFF_FLAGGED(k, EAffect::kGroup)
-			&& k != ch
-			&& !ignores(k, ch, EIgnore::kGroup)) {
-			act(buf, false, ch, nullptr, k, kToVict | kToSleep | kToNotDeaf);
-			// added by WorM  групптелы 2010.10.13
-			if (!AFF_FLAGGED(k, EAffect::kDeafness)
-				&& k->GetPosition() > EPosition::kDead) {
-				sprintf(buf1,
-						"%s сообщил%s группе : '%s'\r\n",
-						tell_can_see(ch, k) ? GET_NAME(ch) : "Кто-то",
-						GET_CH_VIS_SUF_1(ch, k),
-						argument);
-				k->remember_add(buf1, Remember::ALL);
-				k->remember_add(buf1, Remember::GROUP);
-			}
-			//end by WorM
-		}
-		for (f = k->followers; f; f = f->next) {
-			if (AFF_FLAGGED(f->follower, EAffect::kGroup)
-				&& (f->follower != ch)
-				&& !ignores(f->follower, ch, EIgnore::kGroup)) {
-				act(buf, false, ch, nullptr, f->follower, kToVict | kToSleep | kToNotDeaf);
-				// added by WorM  групптелы 2010.10.13
-				if (!AFF_FLAGGED(f->follower, EAffect::kDeafness)
-					&& f->follower->GetPosition() > EPosition::kDead) {
-					sprintf(buf1,
-							"%s сообщил%s группе : '%s'\r\n",
-							tell_can_see(ch, f->follower) ? GET_NAME(ch) : "Кто-то",
-							GET_CH_VIS_SUF_1(ch, f->follower),
-							argument);
-					f->follower->remember_add(buf1, Remember::ALL);
-					f->follower->remember_add(buf1, Remember::GROUP);
-				}
-				//end by WorM
-			}
-		}
-
-		if (ch->IsFlagged(EPrf::kNoRepeat))
-			SendMsgToChar(OK, ch);
-		else {
-			sprintf(buf, "Вы сообщили группе : '%s'\r\n", argument);
-			SendMsgToChar(buf, ch);
-			// added by WorM  групптелы 2010.10.13
-			ch->remember_add(buf, Remember::ALL);
-			ch->remember_add(buf, Remember::GROUP);
-			//end by WorM
-		}
-	}
-}
-
-bool tell_can_see(CharData *ch, CharData *vict) {
-	if (CAN_SEE_CHAR(vict, ch) || IS_IMMORTAL(ch) || GET_INVIS_LEV(ch)) {
-		return true;
-	} else {
-		return false;
-	}
-}
-
 void perform_tell(CharData *ch, CharData *vict, char *arg) {
 	if (vict->IsFlagged(EPrf::kNoInvistell)
 		&& !CAN_SEE(vict, ch)
@@ -413,98 +280,6 @@ void do_spec_comm(CharData *ch, char *argument, int/* cmd*/, int subcmd) {
 		}
 
 		act(action_others, false, ch, nullptr, vict, kToNotVict);
-	}
-}
-
-#define MAX_NOTE_LENGTH 4096    // arbitrary
-
-void do_write(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-	ObjData *paper, *pen = nullptr;
-	char papername[kMaxInputLength], penname[kMaxInputLength];
-
-	two_arguments(argument, papername, penname);
-
-	if (!ch->desc)
-		return;
-
-	if (!*papername)    // nothing was delivered
-	{
-		SendMsgToChar("Написать?  Чем?  И на чем?\r\n", ch);
-		return;
-	}
-	if (*penname)        // there were two arguments
-	{
-		if (!(paper = get_obj_in_list_vis(ch, papername, ch->carrying))) {
-			sprintf(buf, "У вас нет %s.\r\n", papername);
-			SendMsgToChar(buf, ch);
-			return;
-		}
-		if (!(pen = get_obj_in_list_vis(ch, penname, ch->carrying))) {
-			sprintf(buf, "У вас нет %s.\r\n", penname);
-			SendMsgToChar(buf, ch);
-			return;
-		}
-	} else        // there was one arg.. let's see what we can find
-	{
-		if (!(paper = get_obj_in_list_vis(ch, papername, ch->carrying))) {
-			sprintf(buf, "Вы не видите %s в инвентаре.\r\n", papername);
-			SendMsgToChar(buf, ch);
-			return;
-		}
-		if (GET_OBJ_TYPE(paper) == EObjType::kPen)    // oops, a pen..
-		{
-			pen = paper;
-			paper = nullptr;
-		} else if (GET_OBJ_TYPE(paper) != EObjType::kNote) {
-			SendMsgToChar("Вы не можете на ЭТОМ писать.\r\n", ch);
-			return;
-		}
-		// One object was found.. now for the other one.
-		if (!GET_EQ(ch, kHold)) {
-			sprintf(buf, "Вам нечем писать!\r\n");
-			SendMsgToChar(buf, ch);
-			return;
-		}
-		if (!CAN_SEE_OBJ(ch, GET_EQ(ch, kHold))) {
-			SendMsgToChar("Вы держите что-то невидимое!  Жаль, но писать этим трудно!!\r\n", ch);
-			return;
-		}
-		if (pen)
-			paper = GET_EQ(ch, kHold);
-		else
-			pen = GET_EQ(ch, kHold);
-	}
-
-
-	// ok.. now let's see what kind of stuff we've found
-	if (GET_OBJ_TYPE(pen) != EObjType::kPen) {
-		act("Вы не умеете писать $o4.", false, ch, pen, 0, kToChar);
-	} else if (GET_OBJ_TYPE(paper) != EObjType::kNote) {
-		act("Вы не можете писать на $o5.", false, ch, paper, 0, kToChar);
-	} else if (!paper->get_action_description().empty()) {
-		SendMsgToChar("Там уже что-то записано.\r\n", ch);
-	} else            // we can write - hooray!
-	{
-		/* this is the PERFECT code example of how to set up:
-		 * a) the text editor with a message already loaed
-		 * b) the abort buffer if the player aborts the message
-		 */
-		ch->desc->backstr = nullptr;
-		SendMsgToChar("Можете писать.  (/s СОХРАНИТЬ ЗАПИСЬ  /h ПОМОЩЬ)\r\n", ch);
-		// ok, here we check for a message ALREADY on the paper
-		if (!paper->get_action_description().empty())    // we str_dup the original text to the descriptors->backstr
-		{
-			ch->desc->backstr = str_dup(paper->get_action_description().c_str());
-			// send to the player what was on the paper (cause this is already
-			// loaded into the editor)
-			SendMsgToChar(paper->get_action_description().c_str(), ch);
-		}
-		act("$n начал$g писать.", true, ch, 0, 0, kToRoom);
-		// assign the descriptor's->str the value of the pointer to the text
-		// pointer so that we can reallocate as needed (hopefully that made
-		// sense :>)
-		const utils::AbstractStringWriter::shared_ptr writer(new CActionDescriptionWriter(*paper));
-		string_write(ch->desc, writer, MAX_NOTE_LENGTH, 0, nullptr);
 	}
 }
 
@@ -960,241 +735,4 @@ void do_offtop(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
 		}
 	}
 	ch->remember_add(buf1, Remember::OFFTOP);
-}
-
-void ignore_usage(CharData *ch) {
-	SendMsgToChar("Формат команды: игнорировать <имя|все> <режим|все> <добавить|убрать>\r\n"
-				  "Доступные режимы:\r\n"
-				  "  сказать говорить шептать спросить эмоция кричать\r\n"
-				  "  болтать орать группа дружина союзники\r\n", ch);
-}
-
-int ign_find_id(char *name, long *id) {
-	for (auto &i : player_table) {
-		if (!str_cmp(name, i.name())) {
-			if (i.level >= kLvlImmortal) {
-				return 0;
-			}
-
-			*id = i.uid();
-			return 1;
-		}
-	}
-	return -1;
-}
-
-const char *ign_find_name(long id) {
-	for (const auto &i : player_table) {
-		if (id == i.uid()) {
-			return i.name();
-		}
-	}
-
-	return "кто-то";
-}
-
-char *text_ignore_modes(unsigned long mode, char *buf) {
-	buf[0] = 0;
-
-	if (IS_SET(mode, EIgnore::kTell))
-		strcat(buf, " сказать");
-	if (IS_SET(mode, EIgnore::kSay))
-		strcat(buf, " говорить");
-	if (IS_SET(mode, EIgnore::kWhisper))
-		strcat(buf, " шептать");
-	if (IS_SET(mode, EIgnore::kAsk))
-		strcat(buf, " спросить");
-	if (IS_SET(mode, EIgnore::kEmote))
-		strcat(buf, " эмоция");
-	if (IS_SET(mode, EIgnore::kShout))
-		strcat(buf, " кричать");
-	if (IS_SET(mode, EIgnore::kGossip))
-		strcat(buf, " болтать");
-	if (IS_SET(mode, EIgnore::kHoller))
-		strcat(buf, " орать");
-	if (IS_SET(mode, EIgnore::kGroup))
-		strcat(buf, " группа");
-	if (IS_SET(mode, EIgnore::kClan))
-		strcat(buf, " дружина");
-	if (IS_SET(mode, EIgnore::kAlliance))
-		strcat(buf, " союзники");
-	if (IS_SET(mode, EIgnore::kOfftop))
-		strcat(buf, " оффтопик");
-	return buf;
-}
-
-void do_ignore(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-	char arg1[kMaxInputLength];
-	char arg2[kMaxInputLength];
-	char arg3[kMaxInputLength];
-	unsigned int mode = 0, list_empty = 1, all = 0, flag = 0;
-	long vict_id;
-	char buf[256], buf1[256], name[50];
-
-	argument = one_argument(argument, arg1);
-	argument = one_argument(argument, arg2);
-	argument = one_argument(argument, arg3);
-
-// криво попросили -- выведем справку
-	if (arg1[0] && (!arg2[0] || !arg3[0])) {
-		ignore_usage(ch);
-		return;
-	}
-// при вызове без параметров выведем весь список
-	if (!arg1[0] && !arg2[0] && !arg3[0]) {
-		sprintf(buf, "%sВы игнорируете следующих персонажей:%s\r\n", kColorWht, kColorNrm);
-		SendMsgToChar(buf, ch);
-		for (const auto &ignore : ch->get_ignores()) {
-			if (!ignore->id)
-				continue;
-			if (ignore->id == -1) {
-				strcpy(name, "Все");
-			} else {
-				strcpy(name, ign_find_name(ignore->id));
-				name[0] = UPPER(name[0]);
-			}
-			sprintf(buf, "  %s: ", name);
-			SendMsgToChar(buf, ch);
-			mode = ignore->mode;
-			SendMsgToChar(text_ignore_modes(mode, buf), ch);
-			SendMsgToChar("\r\n", ch);
-			list_empty = 0;
-		}
-
-		if (list_empty) {
-			SendMsgToChar("  список пуст.\r\n", ch);
-		}
-		return;
-	}
-
-	if (utils::IsAbbr(arg2, "все"))
-		all = 1;
-	else if (utils::IsAbbr(arg2, "сказать"))
-		flag = EIgnore::kTell;
-	else if (utils::IsAbbr(arg2, "говорить"))
-		flag = EIgnore::kSay;
-	else if (utils::IsAbbr(arg2, "шептать"))
-		flag = EIgnore::kWhisper;
-	else if (utils::IsAbbr(arg2, "спросить"))
-		flag = EIgnore::kAsk;
-	else if (utils::IsAbbr(arg2, "эмоция"))
-		flag = EIgnore::kEmote;
-	else if (utils::IsAbbr(arg2, "кричать"))
-		flag = EIgnore::kShout;
-	else if (utils::IsAbbr(arg2, "болтать"))
-		flag = EIgnore::kGossip;
-	else if (utils::IsAbbr(arg2, "орать"))
-		flag = EIgnore::kHoller;
-	else if (utils::IsAbbr(arg2, "группа"))
-		flag = EIgnore::kGroup;
-	else if (utils::IsAbbr(arg2, "дружина"))
-		flag = EIgnore::kClan;
-	else if (utils::IsAbbr(arg2, "союзники"))
-		flag = EIgnore::kAlliance;
-	else if (utils::IsAbbr(arg2, "оффтоп"))
-		flag = EIgnore::kOfftop;
-	else {
-		ignore_usage(ch);
-		return;
-	}
-
-// имени "все" соответствует id -1
-	if (utils::IsAbbr(arg1, "все")) {
-		vict_id = -1;
-	} else {
-		// убедимся, что добавляемый чар на данный момент существует
-		// и он не имм, а заодно получим его id
-		switch (ign_find_id(arg1, &vict_id)) {
-			case 0: SendMsgToChar("Нельзя игнорировать Богов, это плохо кончится.\r\n", ch);
-				return;
-			case -1: SendMsgToChar("Нет такого персонажа, уточните имя.\r\n", ch);
-				return;
-		}
-	}
-
-// ищем жертву в списке
-	ignore_data::shared_ptr ignore = nullptr;
-	for (const auto &ignore_i : ch->get_ignores()) {
-		if (ignore_i->id == vict_id) {
-			ignore = ignore_i;
-			break;
-		}
-	}
-
-	if (utils::IsAbbr(arg3, "добавить")) {
-// создаем новый элемент списка в хвосте, если не нашли
-		if (!ignore) {
-			const auto cur = std::make_shared<ignore_data>();
-			cur->id = vict_id;
-			cur->mode = 0;
-			ch->add_ignore(cur);
-			ignore = cur;
-		}
-		mode = ignore->mode;
-		if (all) {
-			SET_BIT(mode, EIgnore::kTell);
-			SET_BIT(mode, EIgnore::kSay);
-			SET_BIT(mode, EIgnore::kWhisper);
-			SET_BIT(mode, EIgnore::kAsk);
-			SET_BIT(mode, EIgnore::kEmote);
-			SET_BIT(mode, EIgnore::kShout);
-			SET_BIT(mode, EIgnore::kGossip);
-			SET_BIT(mode, EIgnore::kHoller);
-			SET_BIT(mode, EIgnore::kGroup);
-			SET_BIT(mode, EIgnore::kClan);
-			SET_BIT(mode, EIgnore::kAlliance);
-			SET_BIT(mode, EIgnore::kOfftop);
-		} else {
-			SET_BIT(mode, flag);
-		}
-		ignore->mode = mode;
-	} else if (utils::IsAbbr(arg3, "убрать")) {
-		if (!ignore || ignore->id != vict_id) {
-			if (vict_id == -1) {
-				SendMsgToChar("Вы и так не игнорируете всех сразу.\r\n", ch);
-			} else {
-				strcpy(name, ign_find_name(vict_id));
-				name[0] = UPPER(name[0]);
-				sprintf(buf,
-						"Вы и так не игнорируете "
-						"персонажа %s%s%s.\r\n", kColorWht, name, kColorNrm);
-				SendMsgToChar(buf, ch);
-			}
-			return;
-		}
-		mode = ignore->mode;
-		if (all) {
-			mode = 0;
-		} else {
-			REMOVE_BIT(mode, flag);
-		}
-		ignore->mode = mode;
-		if (!mode)
-			ignore->id = 0;
-	} else {
-		ignore_usage(ch);
-		return;
-	}
-	if (mode) {
-		if (ignore->id == -1) {
-			sprintf(buf, "Для всех сразу вы игнорируете:%s.\r\n", text_ignore_modes(ignore->mode, buf1));
-			SendMsgToChar(buf, ch);
-		} else {
-			strcpy(name, ign_find_name(ignore->id));
-			name[0] = UPPER(name[0]);
-			sprintf(buf, "Для персонажа %s%s%s вы игнорируете:%s.\r\n",
-					kColorWht, name, kColorNrm, text_ignore_modes(ignore->mode, buf1));
-			SendMsgToChar(buf, ch);
-		}
-	} else {
-		if (vict_id == -1) {
-			SendMsgToChar("Вы больше не игнорируете всех сразу.\r\n", ch);
-		} else {
-			strcpy(name, ign_find_name(vict_id));
-			name[0] = UPPER(name[0]);
-			sprintf(buf, "Вы больше не игнорируете персонажа %s%s%s.\r\n",
-					kColorWht, name, kColorNrm);
-			SendMsgToChar(buf, ch);
-		}
-	}
 }
